pasportSet/requirements.txt
asgiref==3.4.1
asyncpg==0.25.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-crispy-forms==1.13.0
django-cte==1.1.5
django-filter==21.1
django-import-export==2.7.1
docutils==0.18.1
et-xmlfile==1.1.0
gitdb==4.0.9
GitPython==3.1.18
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
jeepney==0.7.1
keyring==23.4.1
lxml==4.8.0
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.1.5
pkginfo==1.8.2
progress==1.6
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
tablib==3.1.0
tqdm==4.64.0
twine==3.8.0
typing_extensions==4.1.1
urllib3==1.26.9
uvloop==0.14.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
